type of index modified

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -7,7 +7,7 @@
  * Return: the bit value or -1 for failure
  */
 
-int get_bit(unsigned long int n, unsigned long int index)
+int get_bit(unsigned long int n, unsigned int index)
 {
 	unsigned long int lsh, result;
